#24 [feat] : 주간 영양섭취 정보 응답 데이터 형식 수정

+++ src/main/java/com/sweetbalance/backend/dto/response/WeeklyInfoDTO.java
@@ -9,8 +9,8 @@ import java.util.List;
 @Getter @Setter @Builder
 public class WeeklyInfoDTO {
     private int intake;
-    private double totalSugar;
+    private int totalSugar;
     private double averageSugar;
-    private double totalCalories;
+    private int totalCalories;
     private List<DailySugarDTO> dailySugar;
 }
